Iteration 1
Got the event markers working, the user can select the markers. When they are in range, a scene is loaded with just a return button for now. When the user is out of range a pop up window tells the user that they are out of range including a return button.